preprocess_data/requirements.txt
numpy
pandas
matplotlib
librosa
scikit-learn
tqdm
python-dotenv
requests
Pillow
